app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0
python-json-logger==2.0.7
requests==2.31.0
